Incorrect data type when changing already set, bigint properties.

+++ Source/ACE.Entity/AceObject.cs
@@ -780,7 +780,7 @@ namespace ACE.Entity
                 }
                 else
                 {
-                    listItem.PropertyValue = (uint)value;
+                    listItem.PropertyValue = (ulong)value;
                 }
             }
             else
